pyproject.toml
[project]
name = "ai-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
]
